requirements.txt
asgiref==3.9.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
curl-cffi==0.13.0
dj-database-url==3.1.0
django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
frozendict==2.4.7
gunicorn==23.0.0
idna==3.11
multitasking==0.0.12
numpy==2.4.1
packaging==25.0
pandas==2.3.3
peewee==3.19.0
platformdirs==4.5.1
protobuf==6.33.3
psycopg2-binary==2.9.11
pycparser==2.23
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.3
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
websockets==16.0
whitenoise==6.11.0
yfinance==1.0
